Use address literal as default EHLO argument
This commit fixes formatting of IPv4 and IPv6 address literals.
Both are enclosed in brackets and IPv6 is prefixed with `IPv6:`.

Default trait is derived for ClientId.

Default ClientId is `[127.0.0.1]` now, because it passes the most strict
Postfix `smtpd_helo_restrictions` checks.

Fallback domain ClientId is replaced with `localhost.localdomain`,
because a domain must be an FQDN according to RFC 5321.